Extract a shared search bar component
